requirements.txt
Appium-Python-Client==2.9.0
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
colorama==0.4.6
h11==0.14.0
idna==3.6
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
pycparser==2.21
PySocks==1.7.1
pytest==7.4.4
selenium==4.9.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
urllib3==1.26.18
wsproto==1.2.0
